ai-deployment/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
torch==2.6.0+cpu
torchvision==0.21.0+cpu
timm==1.0.15
Pillow==11.2.1
opencv-python-headless==4.11.0.86
numpy==2.2.5
